pyproject.toml
[tool.poetry]
name = "gpt-gh-summariser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gpt_gh_summariser"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.4"
langchain = "^0.0.134"
requests = "^2.28.2"
pygithub = "^1.58.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.2.2"
mypy = "^1.2.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
